Input separated
